SEQUOIADBMAINSTREAM-4807 【BUG】【new】【story】【S3分段上传对象】并发多线程上传多个分段失败，查看db端返回-319错误
git-svn-id: http://192.168.20.11/sequoiadb/trunk@42855 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/tools/sequoias3/src/main/java/com/sequoias3/taskmanager/UploadScanClean.java
@@ -47,7 +47,7 @@ public class UploadScanClean {
 
     @Scheduled(initialDelay = 1000 * 60, fixedDelay = MINUTE)
     public void uploadScanClean(){
-        logger.debug("upload clean scan begin.");
+//        logger.debug("upload clean scan begin.");
         QueryDbCursor invalidUploads = null;
         try {
             invalidUploads = uploadDao.queryInvalidUploads();
@@ -69,8 +69,10 @@ public class UploadScanClean {
                                         if (!uploadMeta.getCsName().equals(part.getCsName())
                                                 || !uploadMeta.getClName().equals(part.getClName())
                                                 || !uploadMeta.getLobId().equals(part.getLobId())) {
-                                            dataDao.deleteObjectDataByLobId(null,
-                                                    part.getCsName(), part.getClName(), part.getLobId());
+                                            if (part.getLobId() != null) {
+                                                dataDao.deleteObjectDataByLobId(null,
+                                                        part.getCsName(), part.getClName(), part.getLobId());
+                                            }
                                         }
                                     }
                                 }else {
@@ -160,7 +162,7 @@ public class UploadScanClean {
                         }
                     }catch (Exception e){
                         transaction.rollback(connectionC);
-                        logger.error("clean invalid complete status failed.. uploadId:"+uploadId, e);
+                        logger.error("clean invalid complete status failed. uploadId:"+uploadId, e);
                     }finally {
                         daoMgr.releaseConnectionDao(connectionC);
                     }
@@ -171,5 +173,6 @@ public class UploadScanClean {
         } finally {
             metaDao.releaseQueryDbCursor(completeStatusCursor);
         }
+//        logger.debug("upload clean scan end.");
     }
 }
